+++ pom.xml
@@ -53,13 +53,6 @@
 	</properties>
 
 	<dependencies>
-		<dependency>
-			<groupId>org.jetbrains</groupId>
-			<artifactId>annotations</artifactId>
-			<version>26.0.2</version>
-			<scope>provided</scope>
-		</dependency>
-
 		<!-- Tests -->
 		<dependency>
 			<groupId>org.junit.jupiter</groupId>
