+++ src/main/java9/module-info.java
@@ -1,5 +1,3 @@
 module org.cryptomator.siv {
-	requires static org.jetbrains.annotations;
-
 	exports org.cryptomator.siv;
 }
